refactor: repository, community 마크다운 렌더링 dynamic import 적용
