+++ src/components/common/MarkdownRenderer.tsx
@@ -1,0 +1,61 @@
+'use client';
+
+import { FC } from 'react';
+import Markdown from 'react-markdown';
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+import { oneLight } from 'react-syntax-highlighter/dist/cjs/styles/prism';
+
+const whiteTheme = {
+  ...oneLight,
+  'pre[class*="language-"]': {
+    ...oneLight['pre[class*="language-"]'],
+    background: '#ffffff',
+  },
+  'code[class*="language-"]': {
+    ...oneLight['code[class*="language-"]'],
+    background: '#ffffff',
+  },
+};
+
+interface Props {
+  content: string;
+}
+
+const MarkdownRenderer: FC<Props> = ({ content }) => {
+  return (
+    <Markdown
+      components={{
+        code({ className, children, ...rest }) {
+          const match = /language-(\w+)/.exec(className || '');
+          if (match) {
+            return (
+              <SyntaxHighlighter
+                style={whiteTheme}
+                language={match[1]}
+                PreTag="div"
+                customStyle={{
+                  padding: '1rem',
+                  borderRadius: '0.5rem',
+                  fontSize: '0.875rem',
+                  backgroundColor: '#ffffff',
+                }}
+              >
+                {String(children).replace(/\n$/, '')}
+              </SyntaxHighlighter>
+            );
+          }
+
+          return (
+            <code className={className} {...rest}>
+              {children}
+            </code>
+          );
+        },
+      }}
+    >
+      {content}
+    </Markdown>
+  );
+};
+
+export default MarkdownRenderer;
