+++ src/components/community/communityDetail/CommunityDetail.tsx
@@ -8,23 +8,14 @@ import Link from 'next/link';
 import Image from 'next/image';
 import useGetAccessToken from '@/hooks/useGetAccessToken';
 import useCheckAccess from '@/hooks/useCheckExistAccess';
-import Markdown from 'react-markdown';
 import './CommunityDetail.scss';
 import { communityKeys } from '@/querykey/community.querykey';
-import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
-import { oneLight } from 'react-syntax-highlighter/dist/cjs/styles/prism';
+import dynamic from 'next/dynamic';
 
-const whiteTheme = {
-  ...oneLight,
-  'pre[class*="language-"]': {
-    ...oneLight['pre[class*="language-"]'],
-    background: '#ffffff',
-  },
-  'code[class*="language-"]': {
-    ...oneLight['code[class*="language-"]'],
-    background: '#ffffff',
-  },
-};
+const MarkdownRenderer = dynamic(() => import('@/components/common/MarkdownRenderer'), {
+  ssr: false,
+  loading: () => <p>로딩 중...</p>,
+});
 
 interface Props {
   onRendered?: () => void;
@@ -53,7 +44,7 @@ interface CommunityDetailResponse {
   data: CommunityDetail;
 }
 
-const CommunityDetailPage = ( { onRendered }: Props ) => {
+const CommunityDetailPage = ({ onRendered }: Props) => {
   const { tilId } = useParams();
   const tilIdNumber = Number(tilId);
   const { callApi } = useFetch();
@@ -119,6 +110,7 @@ const CommunityDetailPage = ( { onRendered }: Props ) => {
 
     return () => cancelAnimationFrame(handle);
   }, [onRendered]);
+
   if (isLoading || !communityDetailData) return;
 
   return (
@@ -168,39 +160,9 @@ const CommunityDetailPage = ( { onRendered }: Props ) => {
           e.preventDefault();
           await navigator.clipboard.writeText(communityDetailData.content);
         }}
+        ref={contentRef}
       >
-        <Markdown
-          components={{
-            code({ className, children, ...rest }) {
-              const match = /language-(\w+)/.exec(className || '');
-              if (match) {
-                return (
-                  <SyntaxHighlighter
-                    style={whiteTheme}
-                    language={match[1]}
-                    PreTag="div"
-                    customStyle={{
-                      padding: '1rem',
-                      borderRadius: '0.5rem',
-                      fontSize: '0.875rem',
-                      backgroundColor: '#ffffff',
-                    }}
-                  >
-                    {String(children).replace(/\n$/, '')}
-                  </SyntaxHighlighter>
-                );
-              }
-
-              return (
-                <code className={className} {...rest}>
-                  {children}
-                </code>
-              );
-            },
-          }}
-        >
-          {communityDetailData.content}
-        </Markdown>
+        <MarkdownRenderer content={communityDetailData.content} />
       </section>
 
       <button
